+++ src/main/java/lotto/domain/lotto/InvalidLottoException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 /**
  * 로또와 관련된 커스텀 예외
