+++ src/main/java/lotto/domain/lottonumber/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottonumber;
 
 /**
  * 로또 번호와 관련된 커스텀 예외
