+++ src/main/java/lotto/domain/result/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import java.util.ArrayList;
 import java.util.Arrays;
