+++ src/test/java/lotto/domain/lotto/LottoFactoryTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lottonumber.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
