+++ src/test/java/lotto/domain/lotto/PaidLottoTest.java
@@ -1,5 +1,8 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lottonumber.BonusLottoNumber;
+import lotto.domain.lottonumber.LottoNumber;
+import lotto.domain.result.Rank;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
