+++ src/test/java/lotto/domain/lottonumber/LottoNumberTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.lottonumber;
 
+import lotto.domain.lottonumber.InvalidLottoNumberException;
+import lotto.domain.lottonumber.LottoNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
