+++ src/test/java/lotto/domain/money/MoneyForLottoTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.money;
 
+import lotto.domain.money.InvalidMoneyForLottoException;
+import lotto.domain.money.MoneyForLotto;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
